Modularize theme functionality
Fixes #13

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/mkgask/simple-design-tool/issues/13?shareId=XXXX-XXXX-XXXX-XXXX).
