+++ src/App.tsx
@@ -1,27 +1,12 @@
 import { Component, createSignal, onCleanup } from 'solid-js';
 import Header from './components/Header';
 import Main from './components/Main';
+import { applyTheme, handleSystemThemeChange } from './modules/Theme';
 
 import 'flowbite';
 
 const App: Component = () => {
   const [theme, setTheme] = createSignal('system');
-
-  const applyTheme = (theme: string) => {
-    document.body.classList.remove('light', 'dark', 'system');
-    if (theme === 'system') {
-      const systemTheme = window.matchMedia('(prefers-color-scheme: dark)').matches ? 'dark' : 'light';
-      document.body.classList.add(systemTheme);
-    } else {
-      document.body.classList.add(theme);
-    }
-  };
-
-  const handleSystemThemeChange = (event: MediaQueryListEvent) => {
-    if (theme() === 'system') {
-      applyTheme(event.matches ? 'dark' : 'light');
-    }
-  };
 
   const systemThemeMediaQuery = window.matchMedia('(prefers-color-scheme: dark)');
   systemThemeMediaQuery.addEventListener('change', handleSystemThemeChange);
